Move `is_nautilus_class` to persistence
