+++ nautilus_trader/backtest/node.py
@@ -24,7 +24,6 @@ from nautilus_trader.backtest.results import BacktestResult
 from nautilus_trader.config.backtest import BacktestDataConfig
 from nautilus_trader.config.backtest import BacktestRunConfig
 from nautilus_trader.config.backtest import BacktestVenueConfig
-from nautilus_trader.core.inspect import is_nautilus_class
 from nautilus_trader.model.currency import Currency
 from nautilus_trader.model.data.bar import Bar
 from nautilus_trader.model.data.base import DataType
@@ -42,6 +41,7 @@ from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.model.orderbook.data import OrderBookDelta
 from nautilus_trader.persistence.batching import batch_files
 from nautilus_trader.persistence.catalog import DataCatalog
+from nautilus_trader.persistence.util import is_nautilus_class
 
 
 class BacktestNode:
